requirements.txt
pygame>=2.0.0
pyinstaller>=5.0.0
screeninfo>=0.8.1
pywin32>=300
Pillow>=9.0.0
